Move files to *IT as IntegrationTest is not a standard failsafe suffix
Changed failsafe configuration, because groups incorrectly picking tests
