+++ extensions/mapstore/pom.xml
@@ -51,27 +51,31 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-shade-plugin</artifactId>
-                <configuration>
-                    <!--
-                    hazelcast-mapstore doesn't have any dependencies which should be included,
-                    no need for jar-with-dependencies
-                    -->
-                    <shadedArtifactAttached>false</shadedArtifactAttached>
-                    <relocations>
-                        <relocation>
+                <executions>
+                    <execution>
+                        <id>jar-with-dependencies</id>
+                        <phase>${shade.jar-with-dependencies.phase}</phase>
+                        <goals>
+                            <goal>shade</goal>
+                        </goals>
+                        <configuration combine.self="override">
                             <!--
-                            We need to relocate our usage of org.apache.calcite to the location used in hazelcast-sql
+                            hazelcast-mapstore doesn't have any dependencies which should be included,
+                            no need for jar-with-dependencies
                             -->
-                            <pattern>org.apache.calcite</pattern>
-                            <shadedPattern>${relocation.root}.org.apache.calcite</shadedPattern>
-                        </relocation>
-                    </relocations>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-failsafe-plugin</artifactId>
-                <version>${maven.failsafe.plugin.version}</version>
+                            <shadedArtifactAttached>false</shadedArtifactAttached>
+                            <relocations>
+                                <relocation>
+                                    <!--
+                                    We need to relocate our usage of org.apache.calcite to the location used in hazelcast-sql
+                                    -->
+                                    <pattern>org.apache.calcite</pattern>
+                                    <shadedPattern>${relocation.root}.org.apache.calcite</shadedPattern>
+                                </relocation>
+                            </relocations>
+                        </configuration>
+                    </execution>
+                </executions>
             </plugin>
         </plugins>
     </build>
