+++ extensions/mapstore/src/test/java/com/hazelcast/mapstore/mariadb/MariaDBGenericMapStoreIT.java
@@ -16,16 +16,16 @@
 
 package com.hazelcast.mapstore.mariadb;
 
-import com.hazelcast.mapstore.GenericMapStoreIntegrationTest;
+import com.hazelcast.mapstore.GenericMapStoreIT;
 import com.hazelcast.test.annotation.NightlyTest;
 import com.hazelcast.test.jdbc.MariaDBDatabaseProvider;
 import org.junit.BeforeClass;
 import org.junit.experimental.categories.Category;
 
 @Category({NightlyTest.class})
-public class MariaDBGenericMapStoreIntegrationTest extends GenericMapStoreIntegrationTest {
+public class MariaDBGenericMapStoreIT extends GenericMapStoreIT {
 
-    public MariaDBGenericMapStoreIntegrationTest() {
+    public MariaDBGenericMapStoreIT() {
         setPrefix("maria_");
     }
